backend/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-migrate==4.0.4
flask-cors==4.0.0
flask-jwt-extended==4.5.2
python-dotenv==1.0.0
mysqlclient==2.2.0
Pillow==10.0.0
torch==2.0.1
transformers==4.31.0
numpy==1.24.3
requests==2.31.0
python-weather==0.1.0
ultralytics==8.0.0  # YOLO 